Move api.ts and helpers.ts to util folder
