+++ src/lib/util/api.ts
@@ -8,7 +8,7 @@ import type {
 } from "@/types";
 import axios from "axios";
 import { get } from "svelte/store";
-import { notify } from "./util/notify";
+import { notify } from "./notify";
 const { MODE } = import.meta.env;
 
 export const baseURL = MODE === "development" ? "http://127.0.0.1:3080" : "/api";
